+++ package.js
@@ -23,21 +23,17 @@ Package.onTest( function( api ){
 });
 
 function configure( api ){
-    const _use = function(){
-        api.use( ...arguments );
-        api.imply( ...arguments );
-    };
-    api.versionsFrom([ '2.9.0', '3.0-rc.0' ]);
-    _use( 'check' );
-    _use( 'blaze-html-templates@2.0.0 || 3.0.0-alpha300.0', 'client' );
-    _use( 'ecmascript' );
-    _use( 'less@4.0.0', 'client' );
-    _use( 'pwix:ui-bootstrap5@2.0.0' );
-    _use( 'random' );
-    _use( 'reactive-var' );
-    _use( 'tmeasday:check-npm-versions@1.0.2 || 2.0.0-beta.0', 'server' );
-    _use( 'tracker' );
-    //_use( 'tracker', 'client' );
+    api.versionsFrom([ '2.9.0', '3.0' ]);
+    api.use( 'check' );
+    api.use( 'blaze-html-templates@2.0.0 || 3.0.0-alpha300.0', 'client' );
+    api.use( 'ecmascript' );
+    api.use( 'less@4.0.0', 'client' );
+    api.use( 'pwix:ui-bootstrap5@2.0.0' );
+    api.use( 'random' );
+    api.use( 'reactive-var' );
+    api.use( 'tmeasday:check-npm-versions@1.0.2 || 2.0.0-beta.0', 'server' );
+    api.use( 'tracker' );
+    //api.use( 'tracker', 'client' );
     api.addFiles( 'src/client/components/Tabbed/Tabbed.js', 'client' );
 }